requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
distlib==0.3.1
Django==3.1
django-bootstrap4==2.2.0
django-countries==7.0
django-richtextfield==1.6
django-summernote==0.8.11.6
djangorestframework==3.12.2
filelock==3.0.12
Pillow==7.2.0
psycopg2==2.8.6
Pygments==2.7.3
pytz==2020.1
reportlab==3.5.48
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
virtualenv==20.0.30
virtualenvwrapper-win==1.2.6
